+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- feat:  add flag `host` in `up` command
+
 ## v2.6.0 - 2025-08-27
 
 - fix: vscode integration in host network mode